pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "osint-tool"
version = "0.1.0"
description = "Unified modular passive-first OSINT framework"
requires-python = ">=3.12"
dependencies = [
    "httpx[socks]>=0.27",
    "aiohttp>=3.9",
    "click>=8.1",
    "rich>=13.0",
    "textual>=0.50",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.2",
    "dnspython>=2.6",
    "python-whois>=0.9",
    "phonenumbers>=8.13",
    "networkx>=3.3",
    "pyvis>=0.3",
    "jinja2>=3.1",
    "folium>=0.16",
    "plotly>=5.20",
    "Pillow>=10.0",
    "imagehash>=4.3",
    "piexif>=1.1",
    "email-validator>=2.1",
    "stem>=1.8",
    "playwright>=1.40",
    "anthropic>=0.25",
    "weasyprint>=61",
]

[project.scripts]
osint = "osint.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["osint"]
